+++ merqube_client_lib/types/__init__.py
@@ -34,6 +34,8 @@ Records = list[dict[str, str]]
 
 TargetPortfoliosDates = list[tuple[pd.Timestamp, EquityBasketPortfolio]]
 
+HTTPResJson = dict[str, Any]
+
 
 @dataclass
 class CreateReturn: