+++ tests/unit/test_merqube_client.py
@@ -1,4 +1,3 @@
-import datetime
 from unittest.mock import MagicMock
 
 import pytest
@@ -20,12 +19,10 @@ def test_single():
     """
 
     class fake_session:
-        call_count = 0
-
         def get_collection_single(self, *args, **kwargs):
-            if self.call_count == 0:
-                self.call_count += 1
-                return {"id": "valid1id"}
+            return manifest
+
+        def get_json(self, *args, **kwargs):
             return manifest
 
     client = MerqubeAPIClientSingleIndex(index_name="valid1", user_session=fake_session())
@@ -39,7 +36,6 @@ def test_single():
         created_by="test@merqube.com",
         last_modified="2023-01-25T22:40:25.552308",
         last_modified_by="test@merqube.com",
-        locked_after=datetime.datetime(2023, 1, 25, 22, 41),
     )
     assert model.administrative == Administrative(role=Role.administration)