+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - 2023-06-28
+- support manifest locking/unlocking
+- some minor breaking changes for index clients (not 1.0.0 yet!) RE positionl vs kwargs - index client functions now accept name or id
+- add lots more testing
+
 ## [0.15.2] - 2023-06-23
 - README update. Need to enhance the circleci logic so we dont have merge job failures for readme updates that dont bump the version